+++ index.html
@@ -257,7 +257,7 @@
 			</div>
 		</section>
 	</template>
-	<script src="js/main.ts"></script>
+	<script type="module" src="js/main.ts"></script>
 </body>
 
 </html>
