+++ js/data.ts
@@ -1,0 +1,32 @@
+const DESCRIPTION_TEXTS = ['отдыхаем на природе с шашлычками',
+	'классно погуляли с собакой, попали под дождь((',
+	'шлю привет из жаркой Испании',
+	'надоело учиться, когда уже каникулы...',
+	'Не важно, кто ты снаружи, главное — кто ты внутри. ©Джейсон Стэтхэм',
+	'Пусть тортик будет самым вкусным,<br>Каждый подарок — самым лучшим,<br>А жизнь прекрасна и ярка.<br>С днем рождения! Ура!'];
+const MESSAGES = ['Всё отлично!',
+	'В целом всё неплохо. Но не всё.',
+	'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+	'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+	'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+	'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const NAMES = [
+	'Данила',
+	'Анастасия',
+	'Мария',
+	'Юрий',
+	'Анна',
+	'Серафима',
+	'Арсений',
+	'Лидия',
+	'Алёна',
+	'Диана',
+	'Михаил',
+	'Варвара',
+	'Ярослав',
+	'Матвей',
+	'Арина'
+];
+
+export {DESCRIPTION_TEXTS, MESSAGES, NAMES};
